+++ components/google-cloud/google_cloud_pipeline_components/proto/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""GCP Launcher Modules for google cloud components."""
+"""GCP resources proto for google cloud components."""
