+++ components/google-cloud/google_cloud_pipeline_components/proto/gcp_resources.proto
