+++ components/google-cloud/tests/experimental/remote/gcp_launcher/test_model_upload_remote_runner.py
@@ -20,7 +20,7 @@ import time
 import unittest
 from unittest import mock
 from google_cloud_pipeline_components.experimental.remote.gcp_launcher import upload_model_remote_runner
-from google_cloud_pipeline_components.experimental.proto.gcp_resources_pb2 import GcpResources
+from google_cloud_pipeline_components.proto.gcp_resources_pb2 import GcpResources
 from google.protobuf import json_format
 import requests
 import google.auth
